# refactor(perf): implement MapFilter manually